[CARE-20] refactor : re-batch package structure #3
