--- src/main/java/com/caring/user_service/common/security/service/UserTokenService.java
@@ -1,10 +1,0 @@
-package com.caring.user_service.common.security.service;
-
-import com.caring.user_service.common.security.dto.JwtToken;
-import org.springframework.security.core.Authentication;
-
-import java.util.Date;
-
-public interface UserTokenService extends TokenService{
-
-}
