+++ src/main/java/com/caring/user_service/presentation/user/controller/UserApiController.java
@@ -1,9 +1,9 @@
-package com.caring.user_service.domain.user.controller;
+package com.caring.user_service.presentation.user.controller;
 
-import com.caring.user_service.domain.user.business.usecase.ReadAllUserUseCase;
-import com.caring.user_service.domain.user.business.usecase.RegisterUserUseCase;
-import com.caring.user_service.domain.user.vo.request.RequestUser;
-import com.caring.user_service.domain.user.vo.response.ResponseUser;
+import com.caring.user_service.presentation.user.usecase.ReadAllUserUseCase;
+import com.caring.user_service.presentation.user.usecase.RegisterUserUseCase;
+import com.caring.user_service.presentation.user.vo.request.RequestUser;
+import com.caring.user_service.presentation.user.vo.response.ResponseUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
