+++ src/main/java/com/caring/user_service/presentation/user/mapper/UserMapper.java
@@ -1,7 +1,7 @@
-package com.caring.user_service.domain.user.mapper;
+package com.caring.user_service.presentation.user.mapper;
 
 import com.caring.user_service.domain.user.dao.entity.User;
-import com.caring.user_service.domain.user.vo.response.ResponseUser;
+import com.caring.user_service.presentation.user.vo.response.ResponseUser;
 import org.mapstruct.Builder;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
