+++ src/main/java/com/caring/user_service/presentation/user/usecase/ReadAllUserUseCase.java
@@ -1,10 +1,10 @@
-package com.caring.user_service.domain.user.business.usecase;
+package com.caring.user_service.presentation.user.usecase;
 
 import com.caring.user_service.common.annotation.UseCase;
 import com.caring.user_service.domain.user.business.adaptor.UserAdaptor;
 import com.caring.user_service.domain.user.dao.entity.User;
-import com.caring.user_service.domain.user.mapper.UserMapper;
-import com.caring.user_service.domain.user.vo.response.ResponseUser;
+import com.caring.user_service.presentation.user.mapper.UserMapper;
+import com.caring.user_service.presentation.user.vo.response.ResponseUser;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
