+++ src/main/java/com/caring/user_service/presentation/user/usecase/RegisterUserUseCase.java
@@ -1,8 +1,8 @@
-package com.caring.user_service.domain.user.business.usecase;
+package com.caring.user_service.presentation.user.usecase;
 
 import com.caring.user_service.common.annotation.UseCase;
 import com.caring.user_service.domain.user.business.domainservice.UserDomainService;
-import com.caring.user_service.domain.user.vo.request.RequestUser;
+import com.caring.user_service.presentation.user.vo.request.RequestUser;
 import lombok.RequiredArgsConstructor;
 
 @UseCase
