+++ src/main/java/com/caring/user_service/presentation/user/vo/response/ResponseUser.java
@@ -1,4 +1,4 @@
-package com.caring.user_service.domain.user.vo.response;
+package com.caring.user_service.presentation.user.vo.response;
 
 import lombok.*;
