+++ src/main/java/com/caring/user_service/security/service/TokenService.java
@@ -1,9 +1,7 @@
-package com.caring.user_service.common.security.service;
+package com.caring.user_service.security.service;
 
-import com.caring.user_service.common.security.dto.JwtToken;
+import com.caring.user_service.security.dto.JwtToken;
 import org.springframework.security.core.Authentication;
-
-import java.util.Date;
 
 public interface TokenService {
     JwtToken login(String userNumber, String password);
