+++ src/main/java/com/caring/user_service/security/service/user/UserTokenService.java
@@ -1,0 +1,7 @@
+package com.caring.user_service.security.service.user;
+
+import com.caring.user_service.security.service.TokenService;
+
+public interface UserTokenService extends TokenService {
+
+}
