+++ src/main/resources/application.yml
@@ -12,6 +12,9 @@ spring:
     username: root
     password: springproject
     driver-class-name: com.mysql.cj.jdbc.Driver
+  redis:
+    host: localhost
+    port: 6379
 #  rabbitmq:
 #    host: 127.0.0.1
 #    port: 5672
